+++ src/index.ts
@@ -1,3 +1,4 @@
 export { DataLogger, dataTracer } from "./logger.js";
 export { LogEvents } from "./log-events.js";
 export { pause } from "./time/pause.js";
+export { calculateBackoff } from "./utils/backoff.js";